feat: add models file
