+++ src/components/models/product.ts
@@ -1,0 +1,13 @@
+export interface Product {
+  id: number,
+  title: string,
+  description: string,
+  price: number,
+  discountPercentage: number,
+  rating: number,
+  stock: number,
+  brand: string,
+  category: string,
+  thumbnail: string,
+  images: string[],
+}
