+++ src/db/products.ts
@@ -1,17 +1,5 @@
 
-export interface Product {
-  id: number,
-  title: string,
-  description: string,
-  price: number,
-  discountPercentage: number,
-  rating: number,
-  stock: number,
-  brand: string,
-  category: string,
-  thumbnail: string,
-  images: string[],
-}
+import { Product } from "../components/models/product";
 
 
 const products: Product[] = [
